#1: extract ArtifactMetadata interface
